Fixed bug

+++ OpenPlottingLibrary/OpenPlottingLibrary/PlaneGenerator.cs
@@ -19,7 +19,7 @@ namespace OpenPlottingLibrary
         {
             List<Point3D> generatedPlane = new List<Point3D>();
           
-            var f = Function.ToFunc<double, float, float>(expr, "x", "y");
+            var f = Function.ToFunc<float, float, float>(expr, "x", "y");
 
             for (float x = xMin; x <= xMax; x += (1/density))
             {
